Changed opto alg, fixed perfect detection, cleanup
Previously, the algorithm checked a random increase, and if it increased loss, substracts that increase without checking if the subtraction decreases loss. Now, it checks both the increase and the decrease and picks the best of either.

For some reason, cutting off a generation that produced a perfect square didn't work, moving the checking and appending spot fixed it.

Cleaned up some superfluous outputs.